+++ src/utils/buttonsHashMap.js
@@ -1,0 +1,24 @@
+const BUTTONS = [
+  { value: "%", role: "special" },
+  { value: "π", role: "special" },
+  { value: "e", role: "special" },
+  { value: "/", role: "operation" },
+  { value: "7", role: "number" },
+  { value: "8", role: "number" },
+  { value: "9", role: "number" },
+  { value: "x", role: "operation" },
+  { value: "4", role: "number" },
+  { value: "5", role: "number" },
+  { value: "6", role: "number" },
+  { value: "-", role: "operation" },
+  { value: "1", role: "number" },
+  { value: "2", role: "number" },
+  { value: "3", role: "number" },
+  { value: "+", role: "operation" },
+  { value: "0", role: "number", long: true },
+  { value: ".", role: "punctuation" },
+  { value: "⌫", role: "erase" },
+  { value: "C", role: "abort", long: true },
+  { value: "=", role: "res" },
+];
+export default BUTTONS;
