+++ src/utils/operations.js
@@ -1,0 +1,3 @@
+const OPERATIONS = ["-", "+", "/", "*"];
+
+export default OPERATIONS;
